Added Confirmation Service

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -15,7 +16,8 @@ export class MessagesComponent implements OnInit {
   pageNumber = 1;
   pageSize = 5;
   loading = false;
-  constructor(private messageService: MessageService) { }
+
+  constructor(private messageService: MessageService,private confirmService:ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -33,9 +35,14 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe({
-      next: () => this.messages?.splice(this.messages.findIndex(m => m.id === id), 1)
+    this.confirmService.confirm('Confirm Delete message','this cannot be undone').subscribe(result=>{
+      if(result){
+        this.messageService.deleteMessage(id).subscribe({
+          next: () => this.messages?.splice(this.messages.findIndex(m => m.id === id), 1)
+        }) 
+      }
     })
+
   }
 
   pageChanged(event: any) {
@@ -46,4 +53,3 @@ export class MessagesComponent implements OnInit {
   }
 
 }
-
